apps/face-api/requirements.txt
# InsightFace Microservice Dependencies
# Lightweight face detection and verification API

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Face detection (InsightFace with ONNX)
insightface>=0.7.3
onnxruntime>=1.16.0

# Note: For GPU support, replace onnxruntime with:
# onnxruntime-gpu>=1.16.0
